Generic Audio BT Device Support
Add support for any Audio Bluetooth Device recognized by the OS. The main aim is to support incoming calls and notifications being read aloud.
The device must be paired with Android OS and be recognized as audio device.

Known Issue: the device card is apparently unresponsive after the first cycle of connection/disconnection.