Reconnection: Fix a bug with the reconnection delay and make it more aggressive
This makes two changes:

First, it fixes a bug where the reconnection delay was always increased before the first reconnection attempt which meant that the first attempt was always 8 seconds later instead of the intended 4 seconds.

Second, it makes the initial delay 125ms instead of 4 seconds, for BLE devices 4 seconds is a pretty huge delay, so starting out more aggressive should give a better experience with the downside of sending 125, 250, 500, 1000, 2000 before hitting the 4 seconds.